% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlr.mcycle-package.R
\docType{package}
\name{sqlr.mcycle-package}
\alias{sqlr.mcycle}
\alias{sqlr.mcycle-package}
\title{sqlr.mcycle: An R Package for the Sleep-Wake and Mestrual Cycle Project}
\description{
Research compendium of the sleep-wake and menstrual cycle 
      project: a systematic quantitative literature review.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gipso.github.io/sqlr.mcycle}
  \item \url{https://gipso.github.io/sqlr.mcycle/}
  \item Report bugs at \url{https://github.com/gipso/sqlr.mcycle/issues/}
}

}
\author{
\strong{Maintainer}: Daniel Vartanian \email{danvartan@gmail.com} (\href{https://orcid.org/0000-0001-7782-759X}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Poliana Silveira da Silva Goncalves \email{polianassgon@gmail.com} (\href{https://orcid.org/0000-0001-7782-759X}{ORCID}) [copyright holder]
  \item Ana Amelia Benedito-Silva \email{aamelia@usp.br} (\href{https://orcid.org/0000-0003-4976-2623}{ORCID}) [scientific advisor]
  \item Mario Pedrazzoli \email{pedrazzo@usp.br} (\href{https://orcid.org/0000-0002-5257-591X}{ORCID}) [scientific advisor]
}

Other contributors:
\itemize{
  \item Interdisciplinary Sleep Research Group (GIPSO) [funder]
  \item University of Sao Paulo (USP) [funder]
}

}
\keyword{internal}
